CODE RUB: Create Exceptions Fondel Inside Assignment Model
