+++ ToDoList.Core.Api.Tests.Unit/Services/Foundations/Assignments/AssignmentServiceTests.Validations.Add.cs
@@ -6,6 +6,7 @@
 using FluentAssertions;
 using Moq;
 using ToDoList.Core.Api.Models.Assignments;
+using ToDoList.Core.Api.Models.Assignments.Exceptions;
 using Xunit;
 
 namespace ToDoList.Core.Api.Tests.Unit.Services.Foundations.Assignments
